backend/requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.6
redis==5.0.8
python-dotenv==1.0.1
websockets==12.0
python-jose[cryptography]